requirements.txt
# 基础依赖
python-dotenv>=0.19.0
pyaudio>=0.2.11
SpeechRecognition>=3.8.1
gTTS>=2.2.3
openai>=0.27.0
pvporcupine>=2.1.0
requests>=2.26.0

# 语音合成
git+https://github.com/SWivid/F5-TTS.git

# 知识库相关
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.19.5

# 网络相关
netifaces>=0.11.0

# Web相关
flask>=2.0.1
flask-socketio>=5.1.1
eventlet>=0.33.0

# 数据库相关
SQLAlchemy>=1.4.0 